+++ packages/learn-card-base/src/components/vcmodal/hooks/useWalletDid.ts
@@ -1,6 +1,8 @@
 import { useState, useEffect, useCallback } from 'react';
 import { BespokeLearnCard } from 'learn-card-base/types/learn-card';
 import useWallet from 'learn-card-base/hooks/useWallet';
+import { getLogger } from '../../../logging/logger';
+const log = getLogger('use-wallet-did');
 
 const useWalletDid = (lc?: BespokeLearnCard) => {
     const [dids, setDids] = useState<string[]>([]);
@@ -15,7 +17,7 @@ const useWalletDid = (lc?: BespokeLearnCard) => {
             // Set the DIDs, ensuring that only unique values are present
             setDids(defaultDid === keyDid ? [defaultDid] : [defaultDid, keyDid]);
         } catch (error) {
-            console.warn('Error initializing wallet DID:', error);
+            log.warn('Error initializing wallet DID:', error);
         }
     }, [lc]);